Add files via upload
The principle of simulation, in the common sense of the term, is to use a model, that is to say an abstract representation of a system or a problem, and to study the evolution of this model without make the real system work. This is for example the case when you go to the bank for a loan: you are told how much you would pay each month if you accept the offer, without necessarily having to do so. Modeling therefore consisted of posing a problem on paper or on a computer in order to study. It is also what is made for the desires, the study of a physical phenomenon, the construction and the development of new planes, etc. In the case of a bank loan, these are relatively basic calculations, but other examples can be used specific methods.

Monte Carlo simulation is a method of estimating a numerical quantity that uses random numbers. Stanisław Ulam and John von Neumann called it that, in reference to gambling in casinos, during the Manhattan Project that produced the first atomic bomb during World War II. The simulation of a bank loan, for example, is not a Monte-Carlo simulation because it involves exact calculations based on the number of monthly payments and the interest rate; no random phenomenon is involved in the calculations.